+++ src/basic/path-util.h
@@ -181,8 +181,6 @@ bool is_device_path(const char *path);
 bool valid_device_node_path(const char *path);
 bool valid_device_allow_pattern(const char *path);
 
-int systemd_installation_has_version(const char *root, unsigned minimal_version);
-
 bool dot_or_dot_dot(const char *path);
 
 static inline const char *skip_dev_prefix(const char *p) {
