+++ src/nspawn/nspawn-util.h
@@ -1,0 +1,4 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+#pragma once
+
+int systemd_installation_has_version(const char *root, unsigned minimal_version);
